Rework base router spec fixtures into payment lifecycle spec
